Moved store into library
