+++ script/core/model/gameStateStore.js
@@ -1,4 +1,4 @@
-import createStore from './createStore';
+import createStore from './../../library/store/createStore';
 import reducer from './reducers';
 
 export default createStore(reducer);
