+++ script/core/model/reducers/index.js
@@ -1,4 +1,4 @@
-import combineReducers from './../combineReducers';
+import combineReducers from './../../../library/store/combineReducers';
 import {
 	ADD_VIEWPORT,
 	CHANGE_VIEWPORT_POSITION,
@@ -7,13 +7,18 @@ import {
 import {
 	ADD_GAME_OBJECT,
 } from './../actions/gameObjects';
+import {
+	ADD_ROOM,
+	ADD_VIEWPORT_TO_ROOM,
+} from './../actions/rooms';
 
 const initialState = {
-	viewports: {},
 	gameObjects: {},
+	rooms: {},
+	viewports: {},
 };
 
-export default combineReducers(viewports, gameObjects);
+export default combineReducers(viewports, gameObjects, rooms);
 
 export function viewports(state = initialState, action) {
 	switch (action.type) {
@@ -50,6 +55,27 @@ export function gameObjects(state = initialState, action) {
 			return Object.assign({}, state, {
 				gameObjects: Object.assign({}, state.gameObjects, {
 					[action.gameObject.id]: action.gameObject,
+				}),
+			});
+		default:
+			return state;
+	}
+}
+
+export function rooms(state = initialState, action) {
+	switch (action.type) {
+		case ADD_ROOM:
+			return Object.assign({}, state, {
+				rooms: Object.assign({}, state.rooms, {
+					[action.room.id]: action.room,
+				}),
+			});
+		case ADD_VIEWPORT_TO_ROOM:
+			return Object.assign({}, state, {
+				rooms: Object.assign({}, state.rooms, {
+					[action.id]: Object.assign({}, state.rooms[action.id], {
+						viewports: [...state.rooms[action.id].viewports, action.viewportId],
+					}),
 				}),
 			});
 		default:
